HBX-1689: Move class 'org.hibernate.tool.ide.formatting.JavaFormatter' to 'org.hibernate.tool.api.java.Formatter'
